backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
bcrypt
PyJWT

